WIP: Social links for multiple Polkadot ecosystem orgs
Too many socials in footer looks really cluttered
One solution might be to have a dropdown where the user can select
the org to see their socials.

<option> elements however can only contain text nodes, so will need
to work on this a little more.

I'm commiting this WIP version so I can come back to it later. But
the next commit will remove the WIP elements and only show the
Polkadot socials for the time being